efivars: drop unused "optimization"
This was added in c242a082793df77a1dc0bce7f470660ab0a86fe5, and AFAICT, the
code was never exercised, not even in the tests. With this chunk gone, if
anyone ever calls the function without any output params, we'll do open + fstat
instead of access, which will work just fine too.